switch to the devteam's package_samtools_0_1_19
The IUC package is failing to install on test.galaxyproject.org. The devteam's package uses pre-compiled binaries and works out of the box. I created an issue for this here: https://github.com/galaxyproject/tools-iuc/issues/60

It might be worth to switch to the devteam's pre-compiled binaries if possible even if the issue above is solved. Hence this PR.

Sources of the packages:
https://github.com/galaxyproject/tools-devteam/blob/master/packages/package_samtools_0_1_19/tool_dependencies.xml
https://github.com/galaxyproject/tools-iuc/blob/master/packages/package_samtools_0_1_19/tool_dependencies.xml